implemented authorization

+++ backend/routes/posts.js
@@ -34,13 +34,12 @@ router.post('', checkAuth, multer({storage: storage}).single('image'),(req, res,
         title: req.body.title,
         content: req.body.content,
         author: req.body.author,
-        creator: 'userId',
+        creator: req.userData.userId,
         date: req.body.date,
         likes: req.body.likes,
         imagePath: url + "/images/" + req.file.filename,
         tags: req.body.tags
     });
-    console.log(post);
     post.save().then((addedPost) => {
         res.status(201).json({
             message: 'Post was posted successfully!',
@@ -98,15 +97,22 @@ router.put('/:id', checkAuth, multer({storage: storage}).single('image'),(req, r
         title: req.body.title,
         content: req.body.content,
         author: req.body.author,
+        creator: req.userData.userId,
         date: req.body.date,
         imagePath: imagePath
     });
-    Post.updateOne({_id: req.params.id}, post).then(result => {
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+        if(result.nModified > 0) {
 
-        res.status(200).json({
-            message: 'Post was updated successfully!',
-            post: post
-        });
+            res.status(200).json({
+                message: 'Post was updated successfully!',
+                post: post
+            });
+        } else {
+            res.status(401).json({
+                message: 'It is not your post!:)',
+            });
+        }
     });
 });
 
@@ -121,8 +127,13 @@ router.get('/:id', (req, res, next) => {
 })
 
 router.delete('/:id', checkAuth,  (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id }).then((result) => {
-        res.status(200).json({ message: 'post deleted' });
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
+        if(result.n > 0) {
+            res.status(200).json({ message: 'post deleted' });
+        } else {
+            res.status(401).json({ message: 'It is not your post!:)' });
+        }
+
     });
 });
